Business Operations Agents/Expense categorization agent/requirements.txt
litellm>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
streamlit>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.0.0
